Single Value Decomposition with runtime window length
* Added Eigen library to do matrix computation.
* Updated ContextWords object define a new file, not list of files
itself. So each ContextWords object is context of a particular file.
* Error Free implement the less memory Word object passing through each
class data member till the SVD method calls. { cost me very much time
and still much to learn in these topic.. }
Also try to make it as API so that with required Header and object file
any one can use it.